+++ lib/kettle/family/member.rb
@@ -1,0 +1,24 @@
+# frozen_string_literal: true
+
+module Kettle
+  module Family
+    Member = Struct.new(
+      :name,
+      :root,
+      :gemspec_path,
+      :version,
+      :dependencies,
+      keyword_init: true
+    ) do
+      def to_h
+        {
+          "name" => name,
+          "root" => root,
+          "gemspec_path" => gemspec_path,
+          "version" => version,
+          "dependencies" => dependencies
+        }
+      end
+    end
+  end
+end